chore: remove old version command
